Improve performance by using a custom causal convolution
tf.atrous_conv2d pads the height dimension of the input to be a multiple
of the dilation rate, which explains why the model was slow and used a
lot of memory.

After switching to the custom causal convolution function in this
commit, performance has increased massively, and it's now possible to
train large models with dilation rates that wrap around multiple times
from 1 to 512.